requirements.txt
langchain
langchain-experimental
langchain-community
langchain-openai
langchain-chroma
langchain-huggingface
openai
streamlit
faiss-cpu
chromadb
python-dotenv
serpapi
plotly
pandas
numpy
matplotlib
seaborn
requests
beautifulsoup4
youtube-transcript-api
PyPDF2
python-docx
pillow
sentence-transformers
tiktoken
sqlalchemy>=2.0.23
duckdb
streamlit-chat
streamlit-agraph
networkx
transformers
torch
scipy>=1.11.0
joblib>=1.3.0
scikit-learn>=1.3.0

# Enterprise Features
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.9.0
redis>=5.0.0
prometheus-client>=0.19.0
python-multipart>=0.0.6
psutil>=5.9.0

# Database
# sqlite3 is built into Python

# Monitoring & Logging
loguru>=0.7.0
structlog>=23.1.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Performance
aiofiles>=23.2.1
asyncio-mqtt>=0.16.1

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
locust>=2.17.0

# Production
gunicorn>=21.2.0
# supervisor>=4.2.5  # Not available on Windows

# AWS Integration
boto3>=1.28.0
botocore>=1.31.0

# Advanced Document Processing
pdfplumber>=0.10.0
pytesseract>=0.3.10
unstructured>=0.10.0

# LLM Fine-Tuning (LoRA, QLoRA, PEFT)
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.14.0

# Advanced Agentic AI
langgraph>=0.0.20
langchain-openai>=0.0.5

# Multi-Cloud Gen AI
google-cloud-aiplatform>=1.38.0
openai>=1.12.0 